migrations/pyproject.toml
[tool.poetry]
name = "starter-api-migrations"
version = "0.1.0"
description = ""
authors = ["Vince Preziose <vpreziose@firstam.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.6.5"
psycopg2-binary = "^2.9.1"
SQLAlchemy-Utils = "^0.37.8"
Faker = "^8.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
